requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.25
uvicorn==0.27.0.post1
pydantic==2.6.1
fastapi==0.109.2